refactor: move pages related components outside of pages directory
